Refactor competencies
