+++ src/main/java/de/tum/in/www1/artemis/domain/competency/CompetencyProgress.java
@@ -1,4 +1,4 @@
-package de.tum.in.www1.artemis.domain;
+package de.tum.in.www1.artemis.domain.competency;
 
 import java.io.Serial;
 import java.io.Serializable;
@@ -14,6 +14,8 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
+
+import de.tum.in.www1.artemis.domain.User;
 
 /**
  * This class models the 'progress' association between a user and a competency.
