+++ src/main/java/de/tum/in/www1/artemis/domain/competency/CompetencyTaxonomy.java
@@ -1,4 +1,4 @@
-package de.tum.in.www1.artemis.domain;
+package de.tum.in.www1.artemis.domain.competency;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
