+++ src/main/java/de/tum/in/www1/artemis/domain/lecture/LectureUnit.java
@@ -13,6 +13,7 @@ import org.hibernate.annotations.CacheConcurrencyStrategy;
 import com.fasterxml.jackson.annotation.*;
 
 import de.tum.in.www1.artemis.domain.*;
+import de.tum.in.www1.artemis.domain.competency.Competency;
 
 @Entity
 @Table(name = "lecture_unit")
