+++ src/main/java/de/tum/in/www1/artemis/web/rest/util/PageUtil.java
@@ -5,10 +5,10 @@ import javax.validation.constraints.NotNull;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 
-import de.tum.in.www1.artemis.domain.Competency;
 import de.tum.in.www1.artemis.domain.Exercise;
 import de.tum.in.www1.artemis.domain.GradingScale;
 import de.tum.in.www1.artemis.domain.Lecture;
+import de.tum.in.www1.artemis.domain.competency.Competency;
 import de.tum.in.www1.artemis.domain.enumeration.SortingOrder;
 import de.tum.in.www1.artemis.domain.exam.Exam;
 import de.tum.in.www1.artemis.web.rest.dto.PageableSearchDTO;
