+++ src/test/java/de/tum/in/www1/artemis/competency/CompetencyUtilService.java
@@ -3,8 +3,8 @@ package de.tum.in.www1.artemis.competency;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import de.tum.in.www1.artemis.domain.Competency;
 import de.tum.in.www1.artemis.domain.Course;
+import de.tum.in.www1.artemis.domain.competency.Competency;
 import de.tum.in.www1.artemis.repository.CompetencyRepository;
 
 /**
